+++ UefiCpuPkg/PiSmmCpuDxeSmm/CpuS3.c
@@ -6,7 +6,7 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/
 
-#include "PiSmmCpuDxeSmm.h"
+#include "PiSmmCpuCommon.h"
 #include <PiPei.h>
 
 BOOLEAN  mRestoreSmmConfigurationInS3 = FALSE;
