+++ UefiCpuPkg/PiSmmCpuDxeSmm/SmmMp.c
@@ -1,13 +1,13 @@
 /** @file
 SMM MP protocol implementation
 
-Copyright (c) 2019, Intel Corporation. All rights reserved.<BR>
+Copyright (c) 2019 - 2024, Intel Corporation. All rights reserved.<BR>
 
 SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/
 
-#include "PiSmmCpuDxeSmm.h"
+#include "PiSmmCpuCommon.h"
 #include "SmmMp.h"
 
 ///
