+++ UefiCpuPkg/PiSmmCpuDxeSmm/SmmMpPerf.c
@@ -1,13 +1,13 @@
 /** @file
 SMM MP perf-logging implementation
 
-Copyright (c) 2023, Intel Corporation. All rights reserved.<BR>
+Copyright (c) 2023 - 2024, Intel Corporation. All rights reserved.<BR>
 
 SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/
 
-#include "PiSmmCpuDxeSmm.h"
+#include "PiSmmCpuCommon.h"
 
 #define  SMM_MP_PERF_PROCEDURE_NAME(procedure)  # procedure
 GLOBAL_REMOVE_IF_UNREFERENCED
