+++ UefiCpuPkg/PiSmmCpuDxeSmm/SmmProfile.c
@@ -8,7 +8,7 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/
 
-#include "PiSmmCpuDxeSmm.h"
+#include "PiSmmCpuCommon.h"
 #include "SmmProfileInternal.h"
 
 UINT32  mSmmProfileCr3;
