+++ UefiCpuPkg/PiSmmCpuDxeSmm/SyncTimer.c
@@ -6,7 +6,7 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/
 
-#include "PiSmmCpuDxeSmm.h"
+#include "PiSmmCpuCommon.h"
 
 UINT64  mTimeoutTicker = 0;
